requirements.txt
transformers>=2.4.0,<2.5.0
allennlp==0.9.1.dev20200218
backpack-for-pytorch==1.1.0
pytest
pylint